Pertemuan_9/webgis-api/auth_utils.py
"""
auth_utils.py
=============
Utilitas autentikasi JWT dan password hashing untuk WebGIS API.
"""

from datetime import datetime, timedelta
from typing import Optional

from fastapi import Depends, HTTPException, status
from fastapi.security import OAuth2PasswordBearer
from jose import JWTError, jwt
import bcrypt
from dotenv import load_dotenv
import os

load_dotenv()

SECRET_KEY = os.getenv("SECRET_KEY", "changeme-secret-key-very-long-random-string")
ALGORITHM = os.getenv("ALGORITHM", "HS256")
ACCESS_TOKEN_EXPIRE_MINUTES = int(os.getenv("ACCESS_TOKEN_EXPIRE_MINUTES", "60"))

# Skema OAuth2 — titik login
oauth2_scheme = OAuth2PasswordBearer(tokenUrl="/api/auth/login")


# ── Password ──────────────────────────────────────────────────────────────────

def hash_password(plain: str) -> str:
    """Mengembalikan hash bcrypt dari password plaintext."""
    return bcrypt.hashpw(plain.encode("utf-8"), bcrypt.gensalt()).decode("utf-8")


def verify_password(plain: str, hashed: str) -> bool:
    """Memverifikasi password plaintext terhadap hash tersimpan."""
    return bcrypt.checkpw(plain.encode("utf-8"), hashed.encode("utf-8"))


# ── JWT Token ─────────────────────────────────────────────────────────────────

def create_access_token(data: dict, expires_delta: Optional[timedelta] = None) -> str:
    """
    Membuat JWT access token.

    Parameters
    ----------
    data : dict
        Payload yang akan di-encode (biasanya berisi `sub` = username).
    expires_delta : timedelta, optional
        Durasi kedaluwarsa; default menggunakan ACCESS_TOKEN_EXPIRE_MINUTES.

    Returns
    -------
    str
        JWT token yang sudah di-encode.
    """
    to_encode = data.copy()
    expire = datetime.utcnow() + (expires_delta or timedelta(minutes=ACCESS_TOKEN_EXPIRE_MINUTES))
    to_encode.update({"exp": expire})
    return jwt.encode(to_encode, SECRET_KEY, algorithm=ALGORITHM)


# ── Dependency: get current user ──────────────────────────────────────────────

async def get_current_user(token: str = Depends(oauth2_scheme)) -> dict:
    """
    FastAPI dependency — memvalidasi JWT dan mengembalikan payload user.
    Gunakan sebagai `Depends(get_current_user)` pada endpoint terproteksi.

    Raises
    ------
    HTTPException 401 jika token tidak valid atau sudah kedaluwarsa.
    """
    credentials_exc = HTTPException(
        status_code=status.HTTP_401_UNAUTHORIZED,
        detail="Token tidak valid atau sudah kedaluwarsa",
        headers={"WWW-Authenticate": "Bearer"},
    )
    try:
        payload = jwt.decode(token, SECRET_KEY, algorithms=[ALGORITHM])
        username: str = payload.get("sub")
        if username is None:
            raise credentials_exc
        return {"username": username}
    except JWTError:
        raise credentials_exc
